[CODE UNIFICATION] Use the AdUnitConfig instead of OriginalAdUnitConfigurationProtocol (#555)
* feat: confirm AdUnitConfig to OriginalAdUnitConfigurationProtocol

* feat: use AdUnitConfig instead of OriginalAdUnitConfigurationProtocol

* feat(test): update tests

* feat: remove original adunit configuration logic
