+++ PrebidMobileTests/RenderingTests/Tests/Prebid/PBMBidRequesterTest.swift
@@ -35,7 +35,7 @@ class PBMBidRequesterTest: XCTestCase {
     
     func testBanner_300x250() {
         let configId = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
-        let adUnitConfig = AdUnitConfig(configID: configId, size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: configId, size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection(onPost: [{ (url, data, timeout, callback) in
             callback(PBMBidResponseTransformer.someValidResponse)
         }])
@@ -59,7 +59,7 @@ class PBMBidRequesterTest: XCTestCase {
     
     func testBanner_invalidAccountID_noRequest() {
         let configId = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
-        let adUnitConfig = AdUnitConfig(configID: configId, size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: configId, size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection()
         let requester = PBMBidRequester(connection: connection,
                                         sdkConfiguration: sdkConfiguration,
@@ -81,7 +81,7 @@ class PBMBidRequesterTest: XCTestCase {
     func testBanner_invalidAccountID_rejectedByServer() {
         let configId = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
         let accountID = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
-        let adUnitConfig = AdUnitConfig(configID: configId, size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: configId, size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection(onPost: [{ (url, data, timeout, callback) in
             callback(PBMBidResponseTransformer.invalidAccountIDResponse(accountID: accountID))
         }])
@@ -103,7 +103,7 @@ class PBMBidRequesterTest: XCTestCase {
     }
     
     func testBanner_invalidConfigID_noRequest() {
-        let adUnitConfig = AdUnitConfig(configID: " \t \t  ", size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: " \t \t  ", size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection()
         let requester = PBMBidRequester(connection: connection,
                                         sdkConfiguration: sdkConfiguration,
@@ -122,7 +122,7 @@ class PBMBidRequesterTest: XCTestCase {
     
     func testBanner_invalidConfigID_rejectedByServer() {
         let configId = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
-        let adUnitConfig = AdUnitConfig(configID: configId, size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: configId, size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection(onPost: [{ (url, data, timeout, callback) in
             callback(PBMBidResponseTransformer.invalidConfigIdResponse(configId: configId))
         }])
@@ -142,7 +142,7 @@ class PBMBidRequesterTest: XCTestCase {
     }
     
     func testBanner_invalidSize() {
-        let adUnitConfig = AdUnitConfig(configID: "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4", size: CGSize(width: -300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4", size: CGSize(width: -300, height: 250))
         let connection = MockServerConnection()
         let requester = PBMBidRequester(connection: connection,
                                         sdkConfiguration: sdkConfiguration,
@@ -160,7 +160,7 @@ class PBMBidRequesterTest: XCTestCase {
     }
     
     func testBanner_invalidAdditionalSize() {
-        let adUnitConfig = AdUnitConfig(configID: "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4", size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4", size: CGSize(width: 300, height: 250))
         adUnitConfig.additionalSizes = [CGSize(width: -320, height: 50)]
         
         let connection = MockServerConnection()
@@ -181,7 +181,7 @@ class PBMBidRequesterTest: XCTestCase {
     
     func testBanner_requestInProgress() {
         let configId = "b6260e2b-bc4c-4d10-bdb5-f7bdd62f5ed4"
-        let adUnitConfig = AdUnitConfig(configID: configId, size: CGSize(width: 300, height: 250))
+        let adUnitConfig = AdUnitConfig(configId: configId, size: CGSize(width: 300, height: 250))
         let connection = MockServerConnection(onPost: [{ (url, data, timeout, callback) in
             DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
                 callback(PBMBidResponseTransformer.someValidResponse)
